backend/requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
sqlalchemy==1.4.35
pydantic==1.9.0
bcrypt==4.0.1
python-jose==3.3.0
PyJWT==2.3.0
python-multipart==0.0.5